+++ hutool-db/src/main/java/cn/hutool/db/nosql/redis/RedisDS.java
@@ -121,6 +121,8 @@ public class RedisDS implements Closeable, Serializable {
 				setting.getInt("connectionTimeout", group, setting.getInt("timeout", group, Protocol.DEFAULT_TIMEOUT)),
 				// 读取数据超时
 				setting.getInt("soTimeout", group, setting.getInt("timeout", group, Protocol.DEFAULT_TIMEOUT)),
+				// 用户名，issue#I8XEQ4
+				setting.getStr("user", group, null),
 				// 密码
 				setting.getStr("password", group, null),
 				// 数据库序号